backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
cbor2==5.4.6
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
crytic-compile==0.3.4
cytoolz==0.12.2
dataclasses-json==0.6.3
distro==1.8.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.1
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.348
langchain-core==0.0.12
langsmith==0.0.69
lru-dict==1.2.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.8
packaging==23.2
parsimonious==0.9.0
prettytable==3.9.0
protobuf==4.24.3
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.3
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
slither-analyzer==0.9.6
sniffio==1.3.0
solc-select==1.0.4
SQLAlchemy==2.0.23
tenacity==8.2.3
toolz==0.12.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.6
web3==6.10.0
websockets==11.0.3
yarl==1.9.2
